befor clean cod

+++ TestProject.Droid/Services/AudioService.cs
@@ -1,12 +1,6 @@
 ﻿using System;
 using System.IO;
-using System.Threading.Tasks;
-using Android;
-using Android.Content;
-using Android.Content.PM;
 using Android.Media;
-using Android.Support.V4.App;
-using Android.Views;
 using TestProject.Core.Interface;
 using TestProject.Core.Models;
 
@@ -14,14 +8,13 @@ namespace TestProject.Droid.Services
 {
     public class AudioService : IAudioService
     {
-        static long Time { get; set; }
-        MediaRecorder _recorder;
-        MediaPlayer _player;
-        string path;
-        Context _context;
-        private bool _checkAudioFile;
-        static readonly int REQUEST_STORAGE = 0;
-        private View _layout;
+
+        private MediaRecorder _recorder;
+        private MediaPlayer _player;
+        private string path;
+        private string _initialpath = Path.Combine(System.Environment.
+               GetFolderPath(System.Environment.
+               SpecialFolder.Personal), "0" + TwitterUserId.Id_User + ".3gpp");
 
 
 
@@ -36,27 +29,17 @@ namespace TestProject.Droid.Services
             };
         }
 
-        public void PermissionsRecording()
-        {
-            if (ActivityCompat.CheckSelfPermission(_context, Manifest.Permission.RecordAudio) == (int)Permission.Granted)
-            {
-            }
-        }
-
         public void StartRecording(int id)
         {
-                path = Path.Combine(System.Environment.
-                GetFolderPath(System.Environment.
-                SpecialFolder.Personal), "0" + TwitterUserId.Id_User + ".3gpp");
             try
             {
-                if (File.Exists(path))
-                    File.Delete(path);
+                if (File.Exists(_initialpath))
+                    File.Delete(_initialpath);
 
                 _recorder.SetAudioSource(AudioSource.Mic);
                 _recorder.SetOutputFormat(OutputFormat.ThreeGpp);
                 _recorder.SetAudioEncoder(AudioEncoder.AmrNb);
-                _recorder.SetOutputFile(path);
+                _recorder.SetOutputFile(_initialpath);
                 _recorder.Prepare();
                 _recorder.Start();
             }
@@ -75,6 +58,7 @@ namespace TestProject.Droid.Services
                 _recorder.Reset();
                 OnRecordHandler();
             }
+
             catch (Exception e)
             {
                 var str = e.Message;
@@ -83,24 +67,21 @@ namespace TestProject.Droid.Services
 
         public async void PlayRecording(int id)
         {
-            path = Path.Combine(System.Environment.
-                GetFolderPath(System.Environment.
-                SpecialFolder.Personal), "0" + TwitterUserId.Id_User + ".3gpp");
-
             try
             {
                 if (_player == null)
                 {
                     _player = new MediaPlayer();
                 }
+
                 else
                 {
                     _player.Reset();
                 }
 
-                if (File.Exists(path))
+                if (File.Exists(_initialpath))
                 {
-                    Java.IO.File file1 = new Java.IO.File(path);
+                    Java.IO.File file1 = new Java.IO.File(_initialpath);
                     Java.IO.FileInputStream fis1 = new Java.IO.FileInputStream(file1);
                     await _player.SetDataSourceAsync(fis1.FD);
                     _player.Prepare();
@@ -122,7 +103,7 @@ namespace TestProject.Droid.Services
 
                     _player.Completion += PlayCompletion;
                 }
-                
+
             }
 
             catch (Exception ex)
@@ -154,31 +135,35 @@ namespace TestProject.Droid.Services
         {
             path = Path.Combine(System.Environment.
                 GetFolderPath(System.Environment.
-                SpecialFolder.Personal), "0" + TwitterUserId.Id_User + ".3gpp");
+                SpecialFolder.Personal), id.ToString() + TwitterUserId.Id_User + ".3gpp");
 
-            File.Move(path, Path.Combine(System.Environment.
-                GetFolderPath(System.Environment.
-                SpecialFolder.Personal), id.ToString() + TwitterUserId.Id_User + ".3gpp"));
+            if (File.Exists(path))
+            {
+                File.Delete(path);
+                File.Move(_initialpath, path);
+            }
+            else
+            {
+                File.Move(_initialpath, path);
+            }
+            
 
-            File.Delete(path);
+            File.Delete(_initialpath);
 
         }
 
         public void DeleteNullFile()
         {
-            path = Path.Combine(System.Environment.
-                GetFolderPath(System.Environment.
-                SpecialFolder.Personal), "0" + TwitterUserId.Id_User + ".3gpp");
 
-            File.Delete(path);
+            File.Delete(_initialpath);
         }
 
         public bool CheckAudioFile(int id)
         {
-          var  path = Path.Combine(System.Environment.
+            path = Path.Combine(System.Environment.
                 GetFolderPath(System.Environment.
                 SpecialFolder.Personal), id.ToString() + TwitterUserId.Id_User + ".3gpp");
-          var result = File.Exists(path);
+            var result = File.Exists(path);
 
             return result;
